app/test_i2c: Fix i2c read error on ls1021a-iot
The length of the register of SGTL5000 which attached on the i2c0 bus on
ls1021a-iot is 2 Bytes.

The i2c chip INA220 attached on the ls1043ardb and ls1046ardb is not a audio
chip.

The INA220 and SGTL5000 registers is 16 Bits. So it is better to access
with the data length is 2.

Signed-off-by: Wenbin Song <wenbin.song@nxp.com>