configs: ls1021a-tsn: enable LPUART
There are 2 LPUART interfaces on the LS1021A-TSN board, ttyLP0 is routed
to the expansion header at the back of the board, and ttyLP3 is routed
to the Arduino header inside it.

This patch enables the fsl_lpuart driver in U-Boot, and the nodes in the
device tree. Enabling the LPUART driver in the SPL seems to make the SPL
image exceed some sort of size and refuse to boot. This is low priority
(I don't care about serial support in the SPL) so remove that for now.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>