pcie: pcie_layerscape_fixup: avoid write to nonexistent PEX_LUT for LS1028A ECAM
The ls_pcie_lut_set_mapping() function sets up SMMU StreamID mappings
for individual PCIe devices/functions via the PEX_LUT register group.

The fixup function was written for the dedicated PCIe controller, but
since then, there have started to appear SoCs with a PCI ECAM (example:
LS1028A for ENETC functions). Obviously, for the ECAM, this register
group does not exist, and writing to it is going to cause an address
fault.

So patch the fixup function to just set up StreamID mappings for the
devices on buses owned by the real Layerscale PCIe driver.

Suggested-by: Alex Marginean <alexandru.marginean@nxp.com>
Signed-off-by: Catalin Horghidan <catalin.horghidan@nxp.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Jerry Huang <jerry.huang@nxp.com>