baremetal: icc: Fix Kconfig option mismatch issue
Fix the issue which the dependency of CMD_ICC in /cmd/Kconfig not
match the option named "BAREMETAL" in /app/Kconfig.

"BAREMETAL" defined in /app/Kconfig, but the "CMD_ICC" option in
/cmd/Kconfig using "baremetal" to match it. This mismatch will
cause that the icc.c could not be compiled. So, this patch modify
the dependency for "CMD_ICC" in /cmd/Kconfig from "baremetal" to
"BAREMETAL".

Signed-off-by: Chunguang Li <chunguang.li_1@nxp.com>