Fix puts issue for the message confusion
Fix the issue that the core id will separated with the
main information, when different cores put the message
to console at same time.

Limitation:
When the message size is greater than printbuffer size,
the message will be put to console by many times. So,
in this case the core id will has the chance that be
separated with the main log information.

The probability of this happening is not high.

Signed-off-by: Chunguang Li <chunguang.li_1@nxp.com>