fdtdec: use full path in fdtdec_get_alias_seq comparison
Currently fdtdec_get_alias_seq() calls fdt_get_name() which returns only
the name of the leaf node. So it needs to also trim the path of the
alias to the leaf name only, leading to imperfect matches.

This means that the following aliases:

    /aliases {
        eth0 = "/dspi@2120000/ethernet-switch@0/ports/port@0";
        eth1 = "/pcie@1f0000000/pci@0,5/ports/port@0";
    };

will make fdtdec_get_alias_seq to return a seq of zero for both aliases,
as the match will only take place on the "port@0" portion.

Fix this by calling fdt_get_path and comparing the full path of the
alias.

Fixes: 5c33c9fdbb3f ("fdt: Add a function to get the alias sequence of a node")
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>