Remove check_cnt for icc send/perf command handle
Remove check_cnt for icc send/perf command, because of when
large data be sent, the retry count will be very huge and
it is hard to be set, and due to set the interrupt in the
check loop, so the data will be processed eventually.

Add setting the interrupt for icc, when the there is no core
handling the ring, and the ring is not empty.

In current case, it will cause that the data has been sent
succed, but on the send side, it will report send failed.

Signed-off-by: Chunguang Li <chunguang.li_1@nxp.com>
Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>