Fix boot up issue caused by gd->arch.resv_ram init
Add judgement for gd->arch.resv_ram initalization in
drame_init_banksize function that the 2nd DDR size
which the resv_ram alloced from must be greater than 0.

The variable gd->arch.resv_ram is used to track the
location of secure memory, and initalized in
dram_init_banksize function.

In the original logic, when the 2nd DDR size >= the
reserved size, using the 2nd DDR allocate the space
for secure memory. In this situation, when the 2nd DDR
size is 0 (the reserved size could be 0 only), so it
will use the 2nd DDR to allocate space for secure
memory, but there is no space in the 2nd DDR actually,
it will cause unknow address accessing exception when
try to access the secure memory.

So in this judgement the 2nd DDR size must be greater
than 0, then when the 2nd DDR size >= the reserved
size, it will alloce space from 2nd DDR for secure
memory.

Signed-off-by: Chunguang Li <chunguang.li_1@nxp.com>