baremetal: imx8mp: increase delay time before kernel boot up
Increase delay time from 2 to 5 sec before kernel boot up
as a workaround of the conflict between kernel and
baremetal net work.

When the net port connected, the ping test will spend much
time try to ping a dummy IP address. When the kernel booting
during this process, it will hang because of the conflict
between kernel and baremetal network.

Signed-off-by: Chunguang Li <chunguang.li_1@nxp.com>